juranal

+++ pages/lesson/ajax.php
@@ -20,12 +20,22 @@ if (isset($_SESSION['user_id'])) {
         );
     } elseif ($mode == 3) {
         $lesson_id = $_POST['lesson_id'];
-        $success = _exec(
+        $too = 0;
+        _selectRowNoParam(
+            "SELECT count(id) FROM att WHERE lessonid = '$lesson_id'",
+            $too
+            );
+        if($too > 0){
+            echo "Устгах боломжгүй байна. Энэ хичээлийг сонгож ирц бүртгэсэн байна.";
+        }
+        else {
+                    $success = _exec(
             "DELETE FROM tlesson WHERE id = ?",
             'i',
             [$lesson_id],
             $count
         );
+        }
     } elseif ($mode == 4) {
         $lesson_id = $_POST['lesson_id'];
         $tuluv = $_POST['tuluv'];
